Declare an object as wide as its widest read, following merges to their readers

+++ tests/corpus/raw-baseline/arm64_O1/crc32_bitwise.c
@@ -5,21 +5,21 @@ uint64_t sym__crc32_bitwise(uint64_t X0_0, uint64_t X1_0)
         return 0;
     } else {
         uint64_t X9_1;
-        uint64_t X8_1;
+        uint32_t X8_1;
         uint32_t tmp_20380_4;
         X9_1 = 0;
         X8_1 = 0xffffffff;
         uint32_t tmp_2a000_2 = 0xedb88320;
         for (; ; ) {
-            uint64_t X11_5;
+            uint32_t X11_5;
             uint8_t tmp_25600_2 = ((uint8_t*)X0_0)[X9_1];
-            X8_1 = (uint64_t)(uint32_t)((uint32_t)X8_1 ^ (uint32_t)tmp_25600_2);
+            X8_1 = (uint32_t)((uint32_t)X8_1 ^ (uint32_t)tmp_25600_2);
             X11_5 = 8;
             for (; ; ) {
                 tmp_20380_4 = (((uint32_t)X8_1 & 1) * 0xffffffff & tmp_2a000_2) ^ (uint32_t)X8_1 >> 1;
-                X8_1 = (uint64_t)(uint32_t)tmp_20380_4;
+                X8_1 = (uint32_t)tmp_20380_4;
                 uint32_t tmp_lane_1000006d0_a_8_1 = (uint32_t)X11_5;
-                X11_5 = (uint64_t)(uint32_t)(tmp_lane_1000006d0_a_8_1 - 1);
+                X11_5 = (uint32_t)(tmp_lane_1000006d0_a_8_1 - 1);
                 if (tmp_lane_1000006d0_a_8_1 == 1) {
                     break;
                 }
