+++ tests/corpus/raw-baseline/arm64_O2/xxhash32.c
@@ -191,10 +191,10 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
             }
         }
         {
-            uint64_t X11_13;
+            uint32_t X11_13;
             uint64_t X13_2;
             uint32_t tmp_2a000_11 = 0xc2b2ae3d;
-            X11_13 = (uint64_t)(uint32_t)((uint32_t)X1_0 + (uint32_t)X11_8);
+            X11_13 = (uint32_t)((uint32_t)X1_0 + (uint32_t)X11_8);
             uint64_t X13_1 = X12_1 + 4;
             if (X13_1 <= X9_1) {
                 uint32_t tmp_2a000_13 = 0x27d4eb2f;
@@ -204,7 +204,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                     X12_1 += 8;
                     uint32_t tmp_24e00_2 = *(uint32_t*)tmp_7400_5;
                     uint32_t tmp_28e80_2 = tmp_24e00_2 * tmp_2a000_11 + (uint32_t)X11_13;
-                    X11_13 = (uint64_t)(uint32_t)((tmp_28e80_2 >> 15 | tmp_28e80_2 << 17) * tmp_2a000_13);
+                    X11_13 = (uint32_t)((tmp_28e80_2 >> 15 | tmp_28e80_2 << 17) * tmp_2a000_13);
                     uint8_t ZR_7 = X12_1 == X9_1;
                     uint8_t CY_7 = X9_1 <= X12_1;
                     X12_1 = X13_2;
@@ -225,7 +225,7 @@ uint32_t sym__xxhash32(uint64_t X0_0, uint64_t X1_0, uint32_t W2_0)
                         X13_2++;
                         uint8_t tmp_25400_2 = *(uint8_t*)tmp_7400_8;
                         uint32_t tmp_28e80_5 = (uint32_t)tmp_25400_2 * tmp_2a000_2 + (uint32_t)X11_13;
-                        X11_13 = (uint64_t)(uint32_t)((tmp_28e80_5 >> 21 | tmp_28e80_5 << 11) * tmp_2a000_16);
+                        X11_13 = (uint32_t)((tmp_28e80_5 >> 21 | tmp_28e80_5 << 11) * tmp_2a000_16);
                         uint8_t TMPZR_18 = X9_3 == 1;
                         X9_3--;
                         if (TMPZR_18) {
